refactor(e2e-tests): put all infra mode e2e tests under the same package
