+++ test/new-e2e/tests/agent-runtimes/BUILD.bazel
@@ -29,12 +29,6 @@ dd_agent_go_test(
         "auth_artifact_win_test.go",
         "forwarder_nss_failover_test.go",
         "hostname_imdsv2_transition_test.go",
-        "infra_basic_common_test.go",
-        "infra_basic_nix_test.go",
-        "infra_basic_win_test.go",
-        "infra_eudm_common_test.go",
-        "infra_eudm_win_test.go",
-        "infra_utils_test.go",
         "ipc_security_nix_test.go",
         "ipc_security_win_test.go",
         "rtloader_multiprocessing_lib_common_test.go",
@@ -54,7 +48,6 @@ dd_agent_go_test(
         "checks/shared-library/files/minimal_conf.yaml",
         "forwarder/testfixtures/config.yaml.tmpl",
         "forwarder/testfixtures/custom_logs.yaml.tmpl",
-        "infra/fixtures/custom_mycheck.py",
         "rtloader/python-check/multi_pid_check.py",
         "rtloader/python-check/multi_pid_check.yaml",
     ],
@@ -69,7 +62,6 @@ dd_agent_go_test(
         "checks/shared-library/files/minimal_conf.yaml",
         "forwarder/testfixtures/config.yaml.tmpl",
         "forwarder/testfixtures/custom_logs.yaml.tmpl",
-        "infra/fixtures/custom_mycheck.py",
         "rtloader/python-check/multi_pid_check.py",
         "rtloader/python-check/multi_pid_check.yaml",
     ],
