+++ test/new-e2e/tests/agent-runtimes/infra-mode/BUILD.bazel
@@ -1,0 +1,34 @@
+load("//bazel/rules/go:dd_agent_go_test.bzl", "dd_agent_go_test")
+
+dd_agent_go_test(
+    name = "infra-mode_test",
+    srcs = [
+        "infra_basic_common_test.go",
+        "infra_basic_nix_test.go",
+        "infra_basic_win_test.go",
+        "infra_ccm_nix_test.go",
+        "infra_eudm_common_test.go",
+        "infra_eudm_win_test.go",
+        "infra_utils_test.go",
+    ],
+    data = [
+        "fixtures/custom_mycheck.py",
+    ],
+    embedsrcs = [
+        "fixtures/custom_mycheck.py",
+    ],
+    deps = [
+        "//test/e2e-framework/components/datadog/agentparams",
+        "//test/e2e-framework/components/os",
+        "//test/e2e-framework/scenarios/aws/ec2",
+        "//test/e2e-framework/testing/e2e",
+        "//test/e2e-framework/testing/environments",
+        "//test/e2e-framework/testing/provisioners/aws/host",
+        "//test/e2e-framework/testing/testcommon/check",
+        "//test/e2e-framework/testing/utils/e2e/client/agentclient",
+        "//test/fakeintake/aggregator",
+        "//test/fakeintake/client",
+        "@com_github_stretchr_testify//assert",
+        "@com_github_stretchr_testify//require",
+    ],
+)
