+++ test/new-e2e/tests/agent-runtimes/infra-mode/infra_basic_common_test.go
@@ -3,8 +3,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package agentruntimes provides e2e tests for infrastructure mode functionality
-package agentruntimes
+// Package inframode provides e2e tests for infrastructure mode functionality (basic,
+// end_user_device, cloud_cost_only)
+package inframode
 
 import (
 	_ "embed"
@@ -22,7 +23,7 @@ import (
 	awshost "github.com/DataDog/datadog-agent/test/e2e-framework/testing/provisioners/aws/host"
 )
 
-//go:embed infra/fixtures/custom_mycheck.py
+//go:embed fixtures/custom_mycheck.py
 var customCheckPython []byte
 
 var (
