+++ test/new-e2e/tests/agent-runtimes/infra-mode/infra_eudm_common_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package agentruntimes
+package inframode
 
 import (
 	"strings"
